refactor: (#87) util에 있던 MonthRange를 converter로 분리한다.
